ofbid/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
anthropic==0.40.0
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.3.0
pdfplumber==0.11.4
python-multipart==0.0.12
aiosqlite==0.20.0
python-docx==1.1.2
